Create `datafusion-physical-optimizer` crate (#11507)
* Create datafusion-physical-optimizer crate

* fmt .md

* Update Cargo.lock in datafusion-cli

* fmt toml and fix doc
