+++ Cargo.toml
@@ -29,6 +29,7 @@ members = [
     "datafusion/optimizer",
     "datafusion/physical-expr-common",
     "datafusion/physical-expr",
+    "datafusion/physical-optimizer",
     "datafusion/physical-plan",
     "datafusion/proto",
     "datafusion/proto/gen",
@@ -97,6 +98,7 @@ datafusion-functions-array = { path = "datafusion/functions-array", version = "4
 datafusion-optimizer = { path = "datafusion/optimizer", version = "40.0.0", default-features = false }
 datafusion-physical-expr = { path = "datafusion/physical-expr", version = "40.0.0", default-features = false }
 datafusion-physical-expr-common = { path = "datafusion/physical-expr-common", version = "40.0.0", default-features = false }
+datafusion-physical-optimizer = { path = "datafusion/physical-optimizer", version = "40.0.0" }
 datafusion-physical-plan = { path = "datafusion/physical-plan", version = "40.0.0" }
 datafusion-proto = { path = "datafusion/proto", version = "40.0.0" }
 datafusion-proto-common = { path = "datafusion/proto-common", version = "40.0.0" }
