+++ datafusion/physical-optimizer/src/lib.rs
@@ -14,32 +14,9 @@
 // KIND, either express or implied.  See the License for the
 // specific language governing permissions and limitations
 // under the License.
+// Make cheap clones clear: https://github.com/apache/datafusion/issues/11143
+#![deny(clippy::clone_on_ref_ptr)]
 
-//! Optimizer that rewrites [`ExecutionPlan`]s.
-//!
-//! These rules take advantage of physical plan properties , such as
-//! "Repartition" or "Sortedness"
-//!
-//! [`ExecutionPlan`]: crate::physical_plan::ExecutionPlan
-pub mod aggregate_statistics;
-pub mod coalesce_batches;
-pub mod combine_partial_final_agg;
-pub mod enforce_distribution;
-pub mod enforce_sorting;
-pub mod join_selection;
-pub mod limited_distinct_aggregation;
-pub mod optimizer;
-pub mod output_requirements;
-pub mod projection_pushdown;
-pub mod pruning;
-pub mod replace_with_order_preserving_variants;
-pub mod sanity_checker;
-mod sort_pushdown;
-pub mod topk_aggregation;
-pub mod update_aggr_exprs;
-mod utils;
+mod optimizer;
 
-#[cfg(test)]
-pub mod test_utils;
-
-pub use datafusion_physical_optimizer::*;
+pub use optimizer::PhysicalOptimizerRule;
